requirements.txt
astroid @ file:///home/conda/feedstock_root/build_artifacts/astroid_1652111881300/work
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1635267974115/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854175163/work
certifi==2021.10.8
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1606236101993/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1649035228992/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1623610058511/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1651017731964/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
isort @ file:///home/conda/feedstock_root/build_artifacts/isort_1636447814597/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1648854389294/work
lazy-object-proxy @ file:///home/conda/feedstock_root/build_artifacts/lazy-object-proxy_1649033177169/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1651609455247/work
mccabe @ file:///home/conda/feedstock_root/build_artifacts/mccabe_1643049622439/work
munkres==1.1.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020413938/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1652814975413/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1645298319244/work
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1633982426610/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pylint @ file:///home/conda/feedstock_root/build_artifacts/pylint_1652792421559/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857275402/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
torch==1.11.0
torchdata @ file:///home/conda/feedstock_root/build_artifacts/torchdata_1648036481779/work
torchmetrics @ file:///home/conda/feedstock_root/build_artifacts/torchmetrics_1651830953557/work
torchvision==0.12.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650370875435/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1649111919534/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1651495238964/work
